Tests: Fix nprocesses applications under test shutdown
The nprocesses test is using the SIGPIPE signal to shutdown the
spawned child applications. This is done in order to silence
the shell message such as this one:

"nprocesses/test_nprocesses: line 79: 26458 Terminated"

This mechanism is broken on systems where the SIGPIPE default handler
is overriden (thus no application shutdown occurs). This could lead
to stale applications polluting the system and causing heratic behavior
in subsequent tests.

This commit add a loop and check that all the spawned applications have
been successfully killed.