Tests: Fix tests first line of output to follow TAP guidelines
From the Test::Harness::TAP documentation [1]:

     The plan tells how many tests will be run, or how many tests have run.
     It's a check that the test file hasn't stopped prematurely. It must
     appear once, whether at the beginning or end of the output.

The TAP plan *must* be the first line of output of any tests. Currently,
most of the tests output their description and then the TAP plan. This can
cause issues while parsing the test output for "strict" TAP parser such
as the one used in the prove tool.

This commit ensure that the first line outputted by any tests is the TAP
plan.